@@ -18,19 +18,19 @@
     "@paypal/react-paypal-js": "^7.8.1",
     "@sanity/client": "^3.3.3",
     "@sanity/image-url": "^1.0.1",
-    "axios": "^0.27.2",
+    "axios": "^1.6.7",
     "bcryptjs": "^2.4.3",
     "js-cookie": "^3.0.1",
-    "jsonwebtoken": "^8.5.1",
-    "next": "12.2.5",
+    "jsonwebtoken": "^9.0.2",
+    "next": "^14.1.0",
     "next-connect": "^0.13.0",
     "notistack": "^2.0.5",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.34.2"
   },
   "devDependencies": {
     "eslint": "8.22.0",
-    "eslint-config-next": "12.2.5"
+    "eslint-config-next": "^14.1.0"
   }
 }
